move $elemMatch to common tests
